Debugging and UI Cleanups
Files and users that have already been added to groups are now only
listed for deletion on the groupProfilePage and groupAdmin page.

The groupAdmin page is now only available to the owner of the group.

Removed some dead code and fixed comments.

The groupProfilePage and groupAdmin page now have links that return to
the previous page.

Separated groups and files on user profile page.

honeycomb.sql matches the schema in use